Don't fail on client for batch for large requests
