+++ Microsoft.Azure.Cosmos/src/Batch/BatchExecUtils.cs
@@ -7,12 +7,9 @@ namespace Microsoft.Azure.Cosmos
     using System;
     using System.Collections.Generic;
     using System.IO;
-    using System.Net;
     using System.Threading;
     using System.Threading.Tasks;
-    using Microsoft.Azure.Cosmos.Routing;
     using Microsoft.Azure.Documents;
-    using Microsoft.Azure.Documents.Routing;
 
     /// <summary>
     /// Util methods for batch requests.
@@ -26,22 +23,15 @@ namespace Microsoft.Azure.Cosmos
         /// Converts a Stream to a Memory{byte} wrapping a byte array honoring a provided maximum length for the returned Memory.
         /// </summary>
         /// <param name="stream">Stream to be converted to bytes.</param>
-        /// <param name="maximumLength">Desired maximum length of the Memory{byte}.</param>
         /// <param name="cancellationToken"><see cref="CancellationToken"/> to cancel the operation.</param>
         /// <returns>A Memory{byte} with length at most maximumLength.</returns>
         /// <remarks>Throws CosmosException if the input stream has more bytes than maximumLength.</remarks>
         public static async Task<Memory<byte>> StreamToMemoryAsync(
             Stream stream,
-            int maximumLength,
             CancellationToken cancellationToken)
         {
             if (stream.CanSeek)
             {
-                if (stream.Length > maximumLength)
-                {
-                    throw new CosmosException(HttpStatusCode.RequestEntityTooLarge, RMResources.RequestTooLarge);
-                }
-
                 // Some derived implementations of MemoryStream (such as versions of RecyclableMemoryStream prior to 1.2.2 that we may be using)
                 // return an incorrect response from TryGetBuffer. Use TryGetBuffer only on the MemoryStream type and not derived types.
                 MemoryStream memStream = stream as MemoryStream;
@@ -74,10 +64,6 @@ namespace Microsoft.Azure.Cosmos
                     while ((count = await stream.ReadAsync(buffer, 0, bufferSize, cancellationToken)) > 0)
                     {
                         sum += count;
-                        if (sum > maximumLength)
-                        {
-                            throw new CosmosException(HttpStatusCode.RequestEntityTooLarge, RMResources.RequestTooLarge);
-                        }
 
 #pragma warning disable VSTHRD103 // Call async methods when in an async method
                         memoryStream.Write(buffer, 0, count);
@@ -89,39 +75,11 @@ namespace Microsoft.Azure.Cosmos
             }
         }
 
-        public static string GetPartitionKeyRangeId(PartitionKeyInternal partitionKeyInternal, PartitionKeyDefinition partitionKeyDefinition, CollectionRoutingMap collectionRoutingMap)
-        {
-            string effectivePartitionKey = partitionKeyInternal.GetEffectivePartitionKeyString(partitionKeyDefinition);
-            return collectionRoutingMap.GetRangeByEffectivePartitionKey(effectivePartitionKey).Id;
-        }
-
-        public static void GetServerRequestLimits(out int maxServerRequestBodyLength, out int maxServerRequestOperationCount)
-        {
-            maxServerRequestBodyLength = Constants.MaxDirectModeBatchRequestBodySizeInBytes;
-            maxServerRequestOperationCount = Constants.MaxOperationsInDirectModeBatchRequest;
-        }
-
-        public static ResponseMessage EnsureValidStream(
-            IReadOnlyList<ItemBatchOperation> operations,
-            RequestOptions batchOptions,
-            int? maxOperationCount = null)
-        {
-            string errorMessage = BatchExecUtils.EnsureValidInternal(operations, batchOptions, maxOperationCount);
-
-            if (errorMessage != null)
-            {
-                return new ResponseMessage(HttpStatusCode.BadRequest, errorMessage: errorMessage);
-            }
-
-            return new ResponseMessage(HttpStatusCode.OK);
-        }
-
         public static void EnsureValid(
             IReadOnlyList<ItemBatchOperation> operations,
-            RequestOptions batchOptions,
-            int? maxOperationCount = null)
+            RequestOptions batchOptions)
         {
-            string errorMessage = BatchExecUtils.EnsureValidInternal(operations, batchOptions, maxOperationCount);
+            string errorMessage = BatchExecUtils.IsValid(operations, batchOptions);
 
             if (errorMessage != null)
             {
@@ -129,21 +87,15 @@ namespace Microsoft.Azure.Cosmos
             }
         }
 
-        private static string EnsureValidInternal(
+        internal static string IsValid(
             IReadOnlyList<ItemBatchOperation> operations,
-            RequestOptions batchOptions,
-            int? maxOperationCount = null)
+            RequestOptions batchOptions)
         {
             string errorMessage = null;
 
             if (operations.Count == 0)
             {
                 errorMessage = ClientResources.BatchNoOperations;
-            }
-
-            if (maxOperationCount.HasValue && operations.Count > maxOperationCount.Value)
-            {
-                errorMessage = ClientResources.BatchTooLarge;
             }
 
             if (errorMessage == null && batchOptions != null)
